requirements.txt
# App V5 依賴列表 - Render 兼容版本
# 核心依賴（必需）
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
requests>=2.31.0
gunicorn>=21.0.0
sqlalchemy>=2.0.21
pyjwt>=2.8.0
firebase-admin>=6.5.0

# 可選依賴
python-dotenv>=1.0.0
pytest>=7.4.0
